+++ fastly/ngwaf/v1/workspaces/alerts/datadog/globals.go
@@ -1,0 +1,4 @@
+package datadog
+
+// IntegrationType is the integration type for datadog workspace alerts.
+const IntegrationType string = "datadog"