+++ fastly/ngwaf/v1/workspaces/alerts/slack/doc.go
@@ -1,0 +1,3 @@
+// Package slack contains API operations to create, delete, get, list and update.
+// Fastly Next-Gen WAF workspace alerts for Slack integrations.
+package slack